fix(modflow): dataframe support was patchy in a few packages (#2136)
* patch a few remaining gaps in support for dataframes
* update mf6_mnw2_tutorial01.py to use dataframe pkg input
* update a test in test_modflow.py to use dataframe pkg input

+++ flopy/utils/util_list.py
@@ -417,7 +417,9 @@ class MfList(DataInterface, DataListInterface):
         self.__vtype[kper] = np.recarray
 
     def __cast_dataframe(self, kper, d):
-        self.__cast_recarray(kper, d.to_records(index=False))
+        self.__cast_recarray(
+            kper, d.to_records(index=False).astype(self.dtype)
+        )
 
     def get_dataframe(self, squeeze=False):
         """
